+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/assets/application/ImportController.java
@@ -30,16 +30,20 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.eclipse.tractusx.traceability.assets.application.importpoc.ImportService;
-import org.eclipse.tractusx.traceability.assets.application.importpoc.validation.ValidJsonFile;
+import org.eclipse.tractusx.traceability.assets.application.importpoc.validation.JsonFileValidator;
 import org.eclipse.tractusx.traceability.common.response.ErrorResponse;
+import org.eclipse.tractusx.traceability.common.response.ValidationResponse;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
+
+import java.util.List;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -50,6 +54,7 @@ import org.springframework.web.multipart.MultipartFile;
 public class ImportController {
 
     private final ImportService importService;
+    private final JsonFileValidator jsonFileValidator;
 
     @Operation(operationId = "importJson",
             summary = "asset upload",
@@ -79,7 +84,6 @@ public class ImportController {
                     content = @Content(
                             mediaType = "application/json",
                             schema = @Schema(implementation = ErrorResponse.class))),
-
             @ApiResponse(
                     responseCode = "403",
                     description = "Forbidden.",
@@ -110,9 +114,16 @@ public class ImportController {
                     content = @Content(
                             mediaType = "application/json",
                             schema = @Schema(implementation = ErrorResponse.class)))})
-    @PostMapping(value = "/import", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-    public ResponseEntity<Void> importJson(@ValidJsonFile @RequestParam("file") MultipartFile file) {
-        importService.importAssets(file);
+
+    @PostMapping(value = "/import", consumes = MediaType.MULTIPART_FORM_DATA_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<ValidationResponse> importJson(@RequestPart("file") MultipartFile file) {
+        List<String> validationResult = jsonFileValidator.isValid(file);
+        if (!validationResult.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body(new ValidationResponse(validationResult));
+
+        }
+//        importService.importAssets(file);
         return ResponseEntity.noContent().build();
     }
 }